requirements.txt
Flask
Flasgger
torch
torchvision
Pillow
python-dotenv
pymongo
Flask
gunicorn
werkzeug
